refactor: extract button logic into mixin
