+++ packages/button/package.json
@@ -32,6 +32,7 @@
     "web-component"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/component-base": "23.1.0-alpha1",
     "@vaadin/vaadin-lumo-styles": "23.1.0-alpha1",
